added 2 macros, move macros and amounts into global sectino - WIP

+++ playground/src/ParameterId.h
@@ -14,7 +14,7 @@ class ParameterId
   bool operator==(const ParameterId &other) const;
   bool operator!=(const ParameterId &other) const;
 
-  int getNumber() const;
+  uint16_t getNumber() const;
   VoiceGroup getVoiceGroup() const;
 
   std::string toString() const
@@ -31,6 +31,6 @@ class ParameterId
   static bool isGlobal(int number);
 
  private:
-  int m_num;
+  uint16_t m_num;
   VoiceGroup m_group;
 };
